n03 complete: 46,052 clouds, 0 quarantined; two G2 findings; start n05's schema
n03 finished the full corpus. 46,052 of 46,052 sampled, nothing quarantined,
5.6 GB. Verifying it against G2 turned up two things, and neither is fixed by
moving a threshold.

- centroid_offset exceeded G2's 1e-5 for 8 assets, and the clouds are fine.
  Averaging 10,000 float32 values accumulates about 1e-5 of error: recomputing
  one failing asset in float64 gives 5.2e-09 against the recorded 1.15e-05. The
  sidecar statistics are now computed in float64. The gate's threshold is not
  the problem and must not be widened -- the measurement was describing the
  summation method rather than the cloud.
- 21 assets have exactly zero variance on one axis, and their meshes are
  1.8e-12 to 1.7e-08 thick on that axis. They are planes, decals and cutouts;
  all 10,000 points are distinct. An absolute variance floor cannot tell a flat
  ASSET from a flattening BUG, and raising it until they pass would blind the
  check to what it exists for. G2's non-degeneracy clause is now relative to
  raw_bbox_extents, with L1-PC-NONDEGENERATE stating it.

Colour over the full corpus: 51.4% texture, 29.4% flat, 19.2% gltf_default,
zero fallback_grey, mean coloured_point_fraction 1.0000.

n04 died once with BrokenProcessPool at asset 400 -- a GL driver can kill a
child outright. One pool per batch now, and a broken pool costs that batch
rather than the corpus; nothing is corrupted because an asset counts as done
only once its sidecar lands. look_at also rejects a non-finite pose rather than
handing NaN to pyrender, which normalises again internally and can take the
process down instead of raising.

Starts n05_annotate's deterministic half: prompt, schema, closed placement
vocabulary, repair-prompt construction. It carries COMPONENT-OF-NODE rather
than IMPLEMENTS-NODE, because it writes none of n05's channels and the marker
would make the declared-writes check pass on a file that annotates nothing.
27 tests, four negative-injected.

1795 structural checks pass.

+++ metafind/data/pointclouds.py
@@ -379,9 +379,16 @@ def process_one(uid: str, glb: Path, out: Path) -> dict:
         # used to score the annotator as if it were a strong one.
         "raw_bbox_extents": [float(v) for v in extents],
         "raw_bbox_volume": float(np.prod(extents)),
-        "centroid_offset": float(np.abs(normed.mean(axis=0)).max()),
-        "max_radius": float(np.sqrt((normed**2).sum(axis=1)).max()),
-        "per_axis_variance": [float(v) for v in normed.var(axis=0)],
+        # float64 for the STATISTICS, even though the cloud is stored float32.
+        # Averaging 10,000 float32 values accumulates about 1e-5 of error, which
+        # is larger than G2's 1e-5 centroid tolerance -- so 8 of 46,052 assets
+        # were recorded as failing a check their data passes. Verified on one:
+        # recomputed in float64 the centroid offset is 5.2e-09, not 1.15e-05.
+        # The gate's threshold is not the problem and must not be widened; the
+        # measurement was reporting the summation method rather than the cloud.
+        "centroid_offset": float(np.abs(normed.astype(np.float64).mean(axis=0)).max()),
+        "max_radius": float(np.sqrt((normed.astype(np.float64) ** 2).sum(axis=1)).max()),
+        "per_axis_variance": [float(v) for v in normed.astype(np.float64).var(axis=0)],
     }
 
     # The completion marker, written last and atomically. Order matters: npz
